app/src/androidTest/java/com/atdev/unittestingpractice/data/local/ShoppingDaoTest.kt
package com.atdev.unittestingpractice.data.local

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import androidx.room.Room
import androidx.test.core.app.ApplicationProvider
import androidx.test.ext.junit.runners.AndroidJUnit4
import androidx.test.filters.SmallTest
import com.atdev.unittestingpractice.util.getOrAwaitValue
import com.google.common.truth.Truth.assertThat
import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.runBlocking
import kotlinx.coroutines.test.runBlockingTest
import org.junit.After
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import org.junit.runner.RunWith

@RunWith(AndroidJUnit4::class)
@ExperimentalCoroutinesApi
@SmallTest
class ShoppingDaoTest {

    @get: Rule
    var instantTaskExecutorRule = InstantTaskExecutorRule()

    private lateinit var database: LocalDatabase
    private lateinit var dao: ShoppingDao

    @Before
    fun setup() {
        database = Room.inMemoryDatabaseBuilder(
            ApplicationProvider.getApplicationContext(),
            LocalDatabase::class.java
        ).allowMainThreadQueries()
            .build()
        dao = database.shoppingDao()
    }

    @After
    fun tearDown() {
        database.close()
    }

    @Test
    fun insertShoppingItem() = runBlockingTest {
        val shoppingItem =
            ShoppingItem(name = "Ahmed", amount = 12.0, price = 2.0, imageUrl = "", id = 1)
        dao.insertShoppingItem(shoppingItem)

        val allShoppingItems = dao.observeAllShoppingItems().getOrAwaitValue()
        assertThat(allShoppingItems).contains(shoppingItem)
    }

    @Test
    fun deleteShoppingItem() = runBlockingTest {
        val shoppingItem =
            ShoppingItem(name = "Ahmed", amount = 12.0, price = 2.0, imageUrl = "", id = 1)
        dao.insertShoppingItem(shoppingItem)
        dao.deleteShoppingItem(shoppingItem)

        val allShoppingItems = dao.observeAllShoppingItems().getOrAwaitValue()
        assertThat(allShoppingItems).doesNotContain(shoppingItem)
    }

    @Test
    fun observeTotalPriceShoppingItem() = runBlockingTest {
        val shoppingItem1 =
            ShoppingItem(name = "Banana", amount = 2.0, price = 2.0, imageUrl = "", id = 1)
        val shoppingItem2 =
            ShoppingItem(name = "apple", amount = 3.0, price = 3.0, imageUrl = "", id = 2)
        val shoppingItem3 =
            ShoppingItem(name = "carbret", amount = 4.0, price = 4.0, imageUrl = "", id = 3)

        dao.insertShoppingItem(shoppingItem1)
        dao.insertShoppingItem(shoppingItem2)
        dao.insertShoppingItem(shoppingItem3)

        val allShoppingItems = dao.observeTotalPrice().getOrAwaitValue()

        assertThat(allShoppingItems).isEqualTo(2 * 2 + 3 * 3 + 4 * 4)
    }
}